Fix: Room Delete Button
